requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
curl_cffi==0.13.0
frozendict==2.4.7
idna==3.11
llvmlite==0.44.0
multitasking==0.0.12
numba==0.61.2
numpy==2.2.6
pandas==3.0.1
pandas-ta==0.4.71b0
peewee==4.0.0
platformdirs==4.9.2
protobuf==6.33.5
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==16.0
yfinance==1.2.0
